% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{article_assignments}
\alias{article_assignments}
\title{article_assignments}
\format{A data frame with 201 observations of 17 elements:
\describe{
  \item{publication}{}
  \item{year.x}{}
  \item{start_page}{}
  \item{assignee}{}
  \item{doi}{}
  \item{comments}{}
  \item{bibtype}{}
  \item{url}{}
  \item{month}{}
  \item{publisher}{}
  \item{volume}{}
  \item{number}{}
  \item{pages}{}
  \item{author}{}
  \item{title}{}
  \item{journal}{}
  \item{editor}{}
}}
\usage{
article_assignments
}
\description{
article_assignments
}
\keyword{datasets}
