% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assigned_article_notes}
\alias{assigned_article_notes}
\title{assigned_article_notes}
\format{A data frame with 199 observations of 6 elements:
\describe{
\item{journal}{Journal name}
\item{year}{Year}
\item{start_page}{Random number used to select articles wherein this number falls}
\item{assignee}{Person assigned to examine article}
\item{doi}{DOI if available}
\item{comments}{Comments regarding the article}
}}
\usage{
assigned_article_notes
}
\description{
assigned_article_notes
}
\keyword{datasets}
