% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi2bib.R
\name{doi2bib}
\alias{doi2bib}
\alias{doi2bib,character-method}
\alias{doi2bib,ANY-method}
\title{Convert doi to bibtex}
\usage{
doi2bib(doi)

\S4method{doi2bib}{character}(doi)
}
\arguments{
\item{doi}{The character string of the doi.}
}
\value{
a bibtex entry as a character string.
}
\description{
Convert a digital object identifier (doi) string into a bibtex entry using
the web service <http://www.doi2bib.org>.
}
\note{
Original script from William Keith Morris,
<https://gist.github.com/wkmor1/ae2cc96acad472d49c35>.  Documented for use
in an R package by Adam H. Sparks
}
\examples{
doi2bib("10.4454/jpp.fa.2012.002")

}
\author{
William Keith Morris <wkmor1@gmail.com> adapted by Adam H. Sparks
}
