% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi2bib.R
\name{doi2bib}
\alias{doi2bib}
\alias{doi2bib,character-method}
\title{Convert DOI to Bibtex}
\usage{
doi2bib(..., file, append = TRUE, quiet = FALSE)

\S4method{doi2bib}{character}(..., file, append = TRUE, quiet = FALSE)
}
\arguments{
\item{...}{One or more \acronym{DOI}s, as \code{character} strings. If
arguments are named, names will replace default citekeys.}

\item{file}{an optional \code{character} string. If used, the bibtex
references are sent to \code{file} rather than being returned.}

\item{append}{\code{logical}. Append results to file?}

\item{quiet}{\code{logical}. By default, bibtex references are printed to the
console. By setting \code{quiet} to \code{TRUE}, this behaviour will be
prevented.}
}
\value{
a \code{list}, returned invisibly, of bibtex references as
  \code{character} strings, as well as writing to file if \code{file} is
  specified.
}
\description{
Convert a digital object identifier (\acronym{DOI}) string into a bibtex
entry using the web service \url{http://www.doi2bib.org}.
}
\section{Functions}{
\itemize{
\item \code{doi2bib,character-method}: Convert DOI to bibtex
}}

\note{
This function is take from the unrelased (on CRAN) \R package,
 \pkg{doi2bib} by William Keith Morris available from GitHub under the MIT
 Licence.  The full package has been forked to Open Plant Pathology's
 GitHub account \url{https://github.com/openplantpathology/doi2bib}.
}
\examples{
doi2bib(Margules2000 = "10.1038/35012251")
doi2bib(Margules2000 = "10.1038/35012251",
        Myers2000    = "10.1038/35002501",
        Moilanen     = "978-0199547777")
}
\references{
\url{https://github.com/wkmor1/doi2bib}
}
\author{
William Keith Morris, \email{wkmor1@gmail.com}
}
