% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_inter-rater_scores_paper.R
\name{import_interrater_scores_paper}
\alias{import_interrater_scores_paper}
\title{Import scoring notes checking inter-rater repeatability for paper publication}
\usage{
import_interrater_scores_paper()
}
\value{
A [tibble::tibble()] object of reproducibility score notes formatted
for use in analysis and manuscript preparation
}
\description{
Imports, formats data into proper types and calculates the final
reproducibility score.
}
\examples{
notes <- import_notes()
pander::pander(head(notes))

}
