% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rrpp.R
\name{import_rrpp}
\alias{import_rrpp}
\title{Import Reproducibility Score Notes}
\usage{
import_rrpp()
}
\value{
A `tibble` object of reproducibility score notes formatted for use in
 analysis and manuscript preparation
}
\description{
Imports, formats data into proper types and calculates the final
 reproducibility score
}
\examples{
import_rrpp()
pander::pander(head(rrpp))
}
