% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reproducibility.in.Plant.Pathology-package.R
\docType{package}
\name{Reproducibility.in.Plant.Pathology-package}
\alias{Reproducibility.in.Plant.Pathology}
\alias{Reproducibility.in.Plant.Pathology-package}
\title{Reproducibility.in.Plant.Pathology: A Systematic Quantitative Review of Articles Providing a Basis for Identifying Reproducible Research Efforts the Field of Plant Pathology}
\description{
A research compendium for the letter to the editor, "Reproducibility in plant pathology: where do we stand and a way forward", published in "xxxxxx".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openplantpathology/Reproducibility_in_Plant_Pathology}
  \item Report bugs at \url{https://github.com/openplantpathology/Reproducibility_in_Plant_Pathology/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Emerson M. Del Ponte \email{edelponte@gmail.com} (\href{https://orcid.org/0000-0003-4398-409X}{ORCID})
  \item Kaique S. Alves \email{kaiquedsalves@gmail.com} (\href{https://orcid.org/0000-0001-9187-0252}{ORCID})
  \item Zachary S. L. Foster \email{zacharyfoster1989@gmail.com} (\href{https://orcid.org/0000-0002-5075-0948}{ORCID})
  \item Niklaus J. Grünwald \email{grunwaln@science.oregonstate.edu} (\href{https://orcid.org/0000-0003-1656-7602}{ORCID})
}

Other contributors:
\itemize{
  \item William Keith Morris \email{wkmor1@gmail.com} (Wrote the doi2bib() used in our analysis and included in /R.) [contributor]
}

}
\keyword{internal}
