% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoring_criteria}
\alias{scoring_criteria}
\title{Scoring criteria}
\format{
A data frame with 19 rows and three variables:
\describe{
  \item{Category}{Scoring criterion category}
  \item{Index Value}{The index score assigned}
  \item{Definition}{Definition of the criteria used to assign an index value
   score for an article in each category}
}
}
\usage{
scoring_criteria
}
\description{
The scoring criteria used for scoring articles on reproducibility.
}
\keyword{datasets}
