% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assigned_article_notes}
\alias{assigned_article_notes}
\title{assigned_article_notes}
\format{A data frame with 199 observations of 6 elements:
\describe{
\item{publication}{}
\item{year}{}
\item{start_page}{}
\item{assignee}{}
\item{doi}{}
\item{comments}{}
}}
\usage{
assigned_article_notes
}
\description{
assigned_article_notes
}
\keyword{datasets}
