% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_journals.R
\name{table_of_journals}
\alias{table_of_journals}
\title{Create Table of Journals Surveyed}
\usage{
table_of_journals()
}
\value{
pander table of journals surveyed
}
\description{
Create Table of Journals Surveyed
}
\examples{

table_of_journals()

}
