% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_journals.R
\name{table_of_journals}
\alias{table_of_journals}
\title{Create Table of Journals Surveyed}
\usage{
table_of_journals(rrpp)
}
\arguments{
\item{rrpp}{`data.frame` of article evaluations}
}
\value{
\pkg{pander} table of journals surveyed
}
\description{
Create Table of Journals Surveyed
}
\examples{
\donttest{
rrpp <- googlesheets4::read_sheet(
"https://docs.google.com/spreadsheets/d/19gXobV4oPZeWZiQJAPNIrmqpfGQtpapXWcSxaXRw1-M/edit#gid=1699540381"
)
table_of_journals(rrpp)
}

}
