% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dia.R
\name{workflow_dia}
\alias{workflow_dia}
\title{Create Workflow Diagram Figure}
\usage{
workflow_dia()
}
\value{
PDF of workflow diagram
}
\description{
Create Workflow Diagram Figure
}
\examples{

workflow_dia()

}
