% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dia.R
\name{workflow_dia}
\alias{workflow_dia}
\title{Create workflow diagram figure}
\usage{
workflow_dia()
}
\value{
PDF of workflow diagram
}
\description{
Custom function to create a workflow diagram figure, 'Figure 1'.
}
\examples{
workflow_dia()

}
